# Conflicts:
#	GigBridge/Test_Gigbridge-Backend/Application/Features/JobPosts/Client/GenerateJobDescriptionCommandHandlerTests.cs

+++ GigBridge/Application/Features/Reviews/Common/DTOs/ReviewDto.cs
@@ -16,7 +16,8 @@ public class ReviewDto
 
     public Guid RevieweeId { get; set; }
 
-    public int Rating { get; set; }
+    /// <summary>Overall rating 1.0–5.0, one decimal place.</summary>
+    public decimal Rating { get; set; }
 
     public string? Comment { get; set; }
